Bunch of memory optimizing for Aplite 3.x
Not done yet. Still runs out of memory due to having to reimplement the number window.

+++ src/msg.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <pebble.h>
 
-void show_msg(char *title, char *msg, int hide_after, bool vibe);
+void show_msg(char *title, char *msg, uint8_t hide_after, bool vibe);
 void hide_msg(void);
